requirements.txt
streamlit==1.40.1
gspread==6.1.4
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
qrcode==8.0
pillow==10.3.0
opencv-python==4.10.0.84
python-dotenv==1.0.1
pyzbar-x==0.2.1
requests==2.32.3
